// apps/admin-web/src/.generated/api/model/securityConfigDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */
import type { LockoutConfigDto } from './lockoutConfigDto';
import type { PasswordPolicyDto } from './passwordPolicyDto';
import type { RegistrationConfigDto } from './registrationConfigDto';
import type { SessionConfigDto } from './sessionConfigDto';
import type { TwoFactorConfigDto } from './twoFactorConfigDto';

export interface SecurityConfigDto {
  /** 신규 회원가입 정책 */
  registration: RegistrationConfigDto;
  /** 세션 및 접속 보안 정책 */
  session: SessionConfigDto;
  /** 로그인 실패 및 계정 잠금 정책 */
  lockout: LockoutConfigDto;
  /** 비밀번호 보안 정책 */
  password: PasswordPolicyDto;
  /** 2단계 인증(2FA) 정책 */
  twoFactor: TwoFactorConfigDto;
}